main-api-building-code/requirements.txt
faiss-cpu==1.11.0
fastapi==0.115.12
google==3.0.0
google-auth==2.40.3
google-genai==1.20.0
httpx==0.28.1
numpy==2.3.0
pydantic==2.11.5
requests==2.32.4
sentence-transformers==4.1.0
SQLAlchemy==2.0.41
tokenizers==0.21.1
transformers==4.52.4
uvicorn==0.34.3
websockets==15.0.1
